+++ Fabric/src/main/java/fuzs/universalenchants/core/FabricAbstractions.java
@@ -1,11 +1,15 @@
 package fuzs.universalenchants.core;
 
 import fuzs.extensibleenums.core.EnumFactories;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Predicate;
 
@@ -19,5 +23,10 @@ public class FabricAbstractions implements CommonAbstractions{
     @Override
     public EnchantmentCategory createEnchantmentCategory(String enumConstantName, Predicate<Item> predicate) {
         return EnumFactories.createEnchantmentCategory(enumConstantName, predicate);
+    }
+
+    @Override
+    public int getMobLootingLevel(Entity target, @Nullable Entity killer, @Nullable DamageSource cause) {
+        return killer instanceof LivingEntity entity ? EnchantmentHelper.getMobLooting(entity) : 0;
     }
 }