+++ Fabric/src/main/java/fuzs/universalenchants/mixin/FabricBowItemMixin.java
@@ -1,8 +1,10 @@
 package fuzs.universalenchants.mixin;
 
 import fuzs.universalenchants.api.event.entity.player.ArrowLooseCallback;
+import fuzs.universalenchants.handler.ItemCompatHandler;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.item.BowItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ProjectileWeaponItem;
@@ -10,6 +12,7 @@ import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
@@ -18,6 +21,13 @@ public abstract class FabricBowItemMixin extends ProjectileWeaponItem {
 
     public FabricBowItemMixin(Properties p_43009_) {
         super(p_43009_);
+    }
+
+    @ModifyVariable(method = "releaseUsing", at = @At("STORE"), ordinal = 0)
+    public AbstractArrow releaseUsing$storeArrow(AbstractArrow arrow, ItemStack stack) {
+        ItemCompatHandler.applyPiercingEnchantment(arrow, stack);
+        ItemCompatHandler.applyLootingEnchantment(arrow, stack);
+        return arrow;
     }
 
     @Inject(method = "releaseUsing", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/BowItem;getPowerForTime(I)F"), cancellable = true, locals = LocalCapture.CAPTURE_FAILHARD)