+++ Forge/src/main/java/fuzs/universalenchants/UniversalEnchantsForge.java
@@ -4,12 +4,18 @@ import fuzs.puzzleslib.core.CoreServices;
 import fuzs.universalenchants.handler.BetterEnchantsHandler;
 import fuzs.universalenchants.handler.ItemCompatHandler;
 import fuzs.universalenchants.init.ForgeModRegistry;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.living.LivingEntityUseItemEvent;
+import net.minecraftforge.event.entity.living.LivingExperienceDropEvent;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
 import net.minecraftforge.event.entity.living.LootingLevelEvent;
 import net.minecraftforge.event.entity.player.ArrowLooseEvent;
+import net.minecraftforge.event.entity.player.ArrowNockEvent;
+import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.level.BlockEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -39,8 +45,20 @@ public class UniversalEnchantsForge {
             itemCompatHandler.onLootingLevel(evt.getEntity(), evt.getDamageSource(), evt.getLootingLevel()).ifPresent(evt::setLootingLevel);
         });
         BetterEnchantsHandler betterEnchantsHandler = new BetterEnchantsHandler();
-        MinecraftForge.EVENT_BUS.addListener(betterEnchantsHandler::onArrowNock);
-        MinecraftForge.EVENT_BUS.addListener(betterEnchantsHandler::onRightClickItem);
+        MinecraftForge.EVENT_BUS.addListener((final ArrowNockEvent evt) -> {
+            InteractionResultHolder<ItemStack> result = betterEnchantsHandler.onArrowNock(evt.getEntity(), evt.getBow(), evt.getLevel(), evt.getHand());
+            if (result.getResult() != InteractionResult.PASS) {
+                evt.setAction(result);
+                // cancelling the event is only used when firing the arrow is supposed to fail
+            }
+        });
+        MinecraftForge.EVENT_BUS.addListener((final PlayerInteractEvent.RightClickItem evt) -> {
+            InteractionResultHolder<ItemStack> result = betterEnchantsHandler.onRightClickItem(evt.getEntity(), evt.getLevel(), evt.getHand());
+            if (result.getResult() != InteractionResult.PASS) {
+                evt.setCancellationResult(result.getResult());
+                evt.setCanceled(true);
+            }
+        });
         MinecraftForge.EVENT_BUS.addListener((final LivingHurtEvent evt) -> {
             betterEnchantsHandler.onLivingHurt(evt.getEntity(), evt.getSource(), evt.getAmount());
         });
@@ -48,6 +66,8 @@ public class UniversalEnchantsForge {
             betterEnchantsHandler.onFarmlandTrample((Level) evt.getLevel(), evt.getPos(), evt.getState(), evt.getFallDistance(), evt.getEntity()).ifPresent(unit -> evt.setCanceled(true));
         });
         // run after other mods had a chance to change looting level
-        MinecraftForge.EVENT_BUS.addListener(EventPriority.LOW, betterEnchantsHandler::onLivingExperienceDrop);
+        MinecraftForge.EVENT_BUS.addListener(EventPriority.LOW, (final LivingExperienceDropEvent evt) -> {
+            betterEnchantsHandler.onLivingExperienceDrop(evt.getEntity(), evt.getAttackingPlayer(), evt.getOriginalExperience(), evt.getDroppedExperience()).ifPresent(evt::setDroppedExperience);
+        });
     }
 }