+++ Forge/src/main/java/fuzs/universalenchants/core/ForgeAbstractions.java
@@ -1,11 +1,15 @@
 package fuzs.universalenchants.core;
 
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ArrowItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraftforge.common.ForgeHooks;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Predicate;
 
@@ -19,5 +23,10 @@ public class ForgeAbstractions implements CommonAbstractions {
     @Override
     public EnchantmentCategory createEnchantmentCategory(String enumConstantName, Predicate<Item> predicate) {
         return EnchantmentCategory.create(enumConstantName, predicate);
+    }
+
+    @Override
+    public int getMobLootingLevel(Entity target, @Nullable Entity killer, @Nullable DamageSource cause) {
+        return ForgeHooks.getLootingLevel(target, killer, cause);
     }
 }