+++ Forge/src/main/java/fuzs/universalenchants/mixin/ForgeBowItemMixin.java
@@ -11,14 +11,15 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.Slice;
 
 @Mixin(BowItem.class)
-public abstract class BowItemMixin extends ProjectileWeaponItem {
+public abstract class ForgeBowItemMixin extends ProjectileWeaponItem {
 
-    public BowItemMixin(Properties p_43009_) {
+    public ForgeBowItemMixin(Properties p_43009_) {
         super(p_43009_);
     }
 
     @ModifyVariable(method = "releaseUsing", at = @At("STORE"), ordinal = 0, slice = @Slice(to = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/BowItem;customArrow(Lnet/minecraft/world/entity/projectile/AbstractArrow;)Lnet/minecraft/world/entity/projectile/AbstractArrow;", remap = false)))
     public AbstractArrow releaseUsing$storeArrow(AbstractArrow arrow, ItemStack stack) {
+        // Forge patches in another method that overrides the arrow again, we only want to apply once
         ItemCompatHandler.applyPiercingEnchantment(arrow, stack);
         ItemCompatHandler.applyLootingEnchantment(arrow, stack);
         return arrow;