+++ CHANGELOG.md
@@ -6,5 +6,9 @@ The format is based on [Keep a Changelog].
 ## [v4.0.0-1.19.1] - 2022-08-16
 - Ported to Minecraft 1.19.1
 - Split into multi-loader project
+### Added
+- Specialized damage and protection enchantments (e.g. smite and blast protection) are now compatible with the respective base enchantment (sharpness and protection)
+### Changed
+- Enchantment compatibility options are now enabled by default
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/